Adiciona tratamento de erros específico por tipo.

+++ src/main/java/com/sistemaguincho/gestaoguincho/dto/GuinchoResponseDTO.java
@@ -3,6 +3,8 @@ package com.sistemaguincho.gestaoguincho.dto;
 import com.sistemaguincho.gestaoguincho.enums.Disponibilidade;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class GuinchoResponseDTO {
 
@@ -10,7 +12,7 @@ public class GuinchoResponseDTO {
     private String modelo;
     private String placa;
     private String tipo;
-    private Double capacidade;
+    private BigDecimal capacidade;
     private Disponibilidade disponibilidade;
 
 }
