+++ src/main/java/com/sistemaguincho/gestaoguincho/dto/ValoresFixosResponse.java
@@ -3,15 +3,17 @@ package com.sistemaguincho.gestaoguincho.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ValoresFixosResponse {
 
-    private Double valorQuilometragemPorKm;
-    private Double valorQuilometragemSaida;
-    private Double valorMotoristaPorKm;
-    private Double valorMotoristaSaida;
-    private Double valorHoraParada;
-    private Double valorHoraTrabalhada;
-    private Double valorDiaria;
+    private BigDecimal  valorQuilometragemPorKm;
+    private BigDecimal  valorQuilometragemSaida;
+    private BigDecimal  valorMotoristaPorKm;
+    private BigDecimal  valorMotoristaSaida;
+    private BigDecimal  valorHoraParada;
+    private BigDecimal  valorHoraTrabalhada;
+    private BigDecimal valorDiaria;
 }
